Fixed PasswordEncoder, JwtEncoder bean not loading
Spring @Configuration annotation is part of the spring core framework. Spring Configuration annotation indicates that the class has @Bean definition methods. So Spring container can process the class and generate Spring Beans to be used in the application. Without it the Beans wont get loaded neither they'll be able to Autowire in other classes.